move components relating to one functoin into subdirectories
