+++ src/components/blog/Blog.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Blog = () => (
+  <span>blog</span>
+);
+
+export default Blog;
